refactor: Goディレクトリ構成をcmd/internal/レイアウトに整理
core/ → internal/core/、cmd/yomite/ → internal/cli/、app.go → internal/gui/ に
再配置し、CLIエントリポイントを cmd/yomite/main.go に移動。
全importパス・Makefile・GoReleaser設定・仕様書を更新。

Closes #69
