+++ internal/cli/run_test.go
@@ -1,4 +1,4 @@
-package yomite
+package cli
 
 import (
 	"bytes"
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/canpok1/yomite/core"
+	"github.com/canpok1/yomite/internal/core"
 )
 
 // mockProvider はテスト用のProviderモック。
