+++ internal/core/document.go
