+++ internal/core/document_test.go
