+++ internal/core/ollama.go
