+++ internal/core/ollama_test.go
