+++ internal/core/simulator.go
