+++ internal/core/simulator_test.go
