+++ internal/core/templates/embed.go
