+++ internal/core/templates/memory_prompt.tmpl
