+++ internal/core/templates/note_prompt.tmpl
